backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangorestframework-simplejwt[token_blacklist]==5.3.1
django-cors-headers==4.3.1
Pillow==10.4.0
reportlab==4.1.0
qrcode==7.4.2
celery==5.3.6
redis==5.0.6
python-dotenv==1.0.1
drf-spectacular==0.27.1
openpyxl==3.1.5
django-filter==24.2
requests==2.31.0
twilio==9.2.3
cryptography>=42.0.0

